rename fire_storage_service.dart -> storage_provider.dart
